+++ apps/batch/api/routers.py
@@ -5,23 +5,17 @@ This module sets up the DRF router with all viewsets for the batch app.
 """
 from rest_framework.routers import DefaultRouter
 
-# Import main viewsets directly from the viewsets.py file
-import importlib.util
-import os
-
-# Load the viewsets.py file directly
-viewsets_path = os.path.join(os.path.dirname(__file__), 'viewsets.py')
-spec = importlib.util.spec_from_file_location("viewsets_module", viewsets_path)
-viewsets_module = importlib.util.module_from_spec(spec)
-spec.loader.exec_module(viewsets_module)
-
-SpeciesViewSet = viewsets_module.SpeciesViewSet
-BatchViewSet = viewsets_module.BatchViewSet
-BatchContainerAssignmentViewSet = viewsets_module.BatchContainerAssignmentViewSet
-BatchCompositionViewSet = viewsets_module.BatchCompositionViewSet
-BatchTransferViewSet = viewsets_module.BatchTransferViewSet
-MortalityEventViewSet = viewsets_module.MortalityEventViewSet
-GrowthSampleViewSet = viewsets_module.GrowthSampleViewSet
+# Import viewsets from modular structure
+from .viewsets import (
+    SpeciesViewSet,
+    LifeCycleStageViewSet,
+    BatchViewSet,
+    BatchContainerAssignmentViewSet,
+    BatchCompositionViewSet,
+    BatchTransferViewSet,
+    MortalityEventViewSet,
+    GrowthSampleViewSet
+)
 from .viewsets.history import (
     BatchHistoryViewSet,
     BatchContainerAssignmentHistoryViewSet,
@@ -33,6 +27,7 @@ from .viewsets.history import (
 # Create a router and register our viewsets with it
 router = DefaultRouter()
 router.register(r'species', SpeciesViewSet, basename='species')
+router.register(r'lifecycle-stages', LifeCycleStageViewSet, basename='lifecycle-stage')
 router.register(r'batches', BatchViewSet, basename='batch')
 router.register(r'container-assignments', BatchContainerAssignmentViewSet, basename='batch-container-assignment')
 router.register(r'batch-compositions', BatchCompositionViewSet, basename='batch-composition')
